ci: passer Windows et WASM en warnings-as-errors
Les cinq tiers de CI portent désormais la porte. Linux et macOS (Metal et
OpenGL) l'avaient depuis août ; Windows et WASM la rejoignent, et les deux
étaient mal diagnostiqués.

Windows. Ce qui a tenu le drapeau à l'écart pendant trois cycles rouges était
un C4244 levé À L'INTÉRIEUR de <xutility>, depuis une instanciation
std::fill<vector_iterator<uint8_t>, int>. La note du job affirmait que POM1
« ne contient aucun fill( », donc que le site d'appel exigeait la trace
d'instanciation et une machine Windows. C'était faux, et c'est ce qui a envoyé
trois tentatives dans le mur : un grep std::fill( les trouve tous.

  std::fill(v.begin(), v.end(), 0)

sur un std::vector<uint8_t> déduit le type de valeur comme int, donc le
rétrécissement se produit sur le `*_First = _Val;` de <xutility> lui-même —
imputé à l'en-tête et jamais à l'appelant, ce qui est exactement pourquoi il se
lit comme introuvable. Six sites (cinq dans Memory.cpp, un dans
Pom1BenchHost.cpp) écrivent maintenant static_cast<uint8_t>(...), le même
idiome que les cinq std::fill_n voisins utilisaient déjà : la passe précédente
avait corrigé ceux-là et manqué ceux-ci.

WASM. Cinq fonctions que seuls les chemins natifs appellent — loadSizeFile,
hexDumpFilter et trois auxiliaires du bundle cc65 — étaient définies sans
condition alors que leurs seuls appelants vivaient dans un #if !POM1_IS_WASM.
Gardes posées sur les définitions. Vérifié localement : emcmake + -Werror
compile propre et produit un POM1.wasm de 3,37 Mo. (Le décompte annoncé était
« deux statiques inutilisées » ; il y en avait cinq.)

mainwindow_lines 16780 -> 16791 : les 11 lignes de gardes. C'est le seul type
de hausse que ce plafond accepte — une qui nomme ce qu'elle achète.

Le tier Windows n'est PAS vérifiable localement (pas de MSVC sous macOS) :
cette branche existe pour que la CI le dise avant que main ne le porte.

125/125 verts en local.

+++ src/Memory.cpp
@@ -809,7 +809,7 @@ void Memory::initMemory(){
         for (int i = 0; i < ramSize * 1024; ++i)
             mem[i] = static_cast<uint8_t>(dist(gen));
     } else {
-        std::fill(mem.begin(), mem.end(), 0);
+        std::fill(mem.begin(), mem.end(), static_cast<uint8_t>(0));
     }
     resetRamWriteTrap();
     markAllPagesDirty();
@@ -1801,7 +1801,7 @@ void Memory::setMicroSDEnabled(bool b)
         // FLAGS runs after MEM — zeroing here would wipe just-restored RAM
         // (e.g. Applesoft Lite at $6000-$7FFF when switching presets).
         if (!snapshotRestoreInProgress) {
-            std::fill(mem.begin() + 0x8000, mem.begin() + 0xA000, 0);
+            std::fill(mem.begin() + 0x8000, mem.begin() + 0xA000, static_cast<uint8_t>(0));
             markPagesDirty(0x8000, 0x2000);
         }
     }
@@ -1834,7 +1834,7 @@ int Memory::loadSDCardRom()
     bool prev = writeInRom;
     writeInRom = true;
     // Clear region first — ROM file (8177 B) may not fill the full 8 KB space
-    std::fill(mem.begin() + 0x8000, mem.begin() + 0xA000, 0);
+    std::fill(mem.begin() + 0x8000, mem.begin() + 0xA000, static_cast<uint8_t>(0));
     int ret = loadROM("sdcard.rom", 0x8000, 0x2000, "SD CARD OS");
     markPagesDirty(0x8000, 0x2000);
     writeInRom = prev;
@@ -1854,7 +1854,7 @@ void Memory::setCFFA1Enabled(bool b)
         // Clear the CFFA1 ROM region. Skip during snapshot restore — same
         // MEM-then-FLAGS ordering hazard as setMicroSDEnabled(false).
         if (!snapshotRestoreInProgress) {
-            std::fill(mem.begin() + 0x9000, mem.begin() + 0xB000, 0);
+            std::fill(mem.begin() + 0x9000, mem.begin() + 0xB000, static_cast<uint8_t>(0));
             markPagesDirty(0x9000, 0x2000);
         }
     }
@@ -1958,7 +1958,7 @@ void Memory::setJukeBoxEnabled(bool b)
         // that guard. Skip during snapshot restore (MEM-then-FLAGS ordering, same
         // as setCFFA1Enabled/setMicroSDEnabled).
         if (!snapshotRestoreInProgress) {
-            std::fill(mem.begin() + 0x4000, mem.begin() + 0xC000, 0);
+            std::fill(mem.begin() + 0x4000, mem.begin() + 0xC000, static_cast<uint8_t>(0));
             markPagesDirty(0x4000, 0x8000);
         }
     }
